requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15

psycopg[binary]>=3.2

python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.5
django-cors-headers>=4.0
dj-database-url
psycopg2-binary
<<<<<<< HEAD
django-filter
=======
django-filter
djangorestframework-simplejwt
>>>>>>> dev
